07-streaming/extras/extras/pyflink/requirements.txt
apache-flink==1.16.0
psycopg2-binary==2.9.1
requests
kafka-python